+++ src/main/java/nova/command/Command.java
@@ -1,0 +1,13 @@
+package nova.command;
+
+/**
+ * Represents a command that can be executed.
+ */
+public interface Command {
+    /**
+     * Executes the command.
+     *
+     * @return true if execution was successful, false otherwise.
+     */
+    boolean execute();
+}
